Hay Order (#39)
* Update CareTaking.lua

Changed coords to 2nd row for easier access to the activities. No problem for spamming enter to quick and changing coords

* Update server.lua

fixed the double money earned after selling the animals. The money will only go to the ledger

* Update Animals.lua

changed the sorting of the animals listed

* Update FeedAnimals.lua

changed the order of the hay being picked of the cart